[FLINK-36369][table] Move deprecated user-visible classes in table-api-java to the legacy package
