+++ flink-table/flink-table-api-java-bridge/src/main/java/org/apache/flink/table/descriptors/RowtimeValidator.java
@@ -22,9 +22,9 @@ import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.api.java.tuple.Tuple2;
 import org.apache.flink.table.api.ValidationException;
 import org.apache.flink.table.legacy.descriptors.Rowtime;
+import org.apache.flink.table.legacy.sources.tsextractors.ExistingField;
+import org.apache.flink.table.legacy.sources.tsextractors.StreamRecordTimestamp;
 import org.apache.flink.table.legacy.sources.tsextractors.TimestampExtractor;
-import org.apache.flink.table.sources.tsextractors.ExistingField;
-import org.apache.flink.table.sources.tsextractors.StreamRecordTimestamp;
 import org.apache.flink.table.sources.wmstrategies.AscendingTimestamps;
 import org.apache.flink.table.sources.wmstrategies.BoundedOutOfOrderTimestamps;
 import org.apache.flink.table.sources.wmstrategies.PreserveWatermarks;
