+++ flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/legacy/sources/tsextractors/ExistingField.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.sources.tsextractors;
+package org.apache.flink.table.legacy.sources.tsextractors;
 
 import org.apache.flink.annotation.PublicEvolving;
 import org.apache.flink.api.common.typeinfo.TypeInformation;
@@ -26,7 +26,6 @@ import org.apache.flink.table.expressions.Expression;
 import org.apache.flink.table.expressions.FieldReferenceExpression;
 import org.apache.flink.table.expressions.ResolvedFieldReference;
 import org.apache.flink.table.legacy.descriptors.Rowtime;
-import org.apache.flink.table.legacy.sources.tsextractors.TimestampExtractor;
 import org.apache.flink.table.types.DataType;
 
 import java.sql.Timestamp;
